# backend/Document/Swagger.yaml
openapi: 3.0.4
info:
  title: National Football League Management API
  description: |-
    This is an API for managing a national football league, built on the OpenAPI 3.0 specification. It provides endpoints for managing team information, players, match schedules, results, and league standings.  
    Learn more about OpenAPI at [https://swagger.io](https://swagger.io). We follow a design-first approach, allowing you to contribute to improving the API, from its definition to its code.  
    Some useful links:  
    - [National Football League API Repository](https://github.com/truonghoangkhiem/QL_Giai_BD_QG_SE104_UIT.git)
  termsOfService: https://swagger.io/terms/
  contact:
    email: truongkhiemhoang@gmail.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.12
...
servers:
  - url: http://localhost:5000
...
paths:
...
  /api/rankings/{seasonid}:
    get:
      tags:
        - ranking
      summary: Get season rankings
      description: Retrieves all rankings for a specific season
      operationId: getSeasonRanking
      parameters:
        - $ref: "#/components/parameters/SeasonId"
      responses:
        "200":
          description: Season rankings fetched successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RankingsResponse"
        "400":
          description: Invalid season ID format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      tags:
        - ranking
      summary: Update season rankings
      description: Updates or creates rankings for a specific season based on a match date
      operationId: updateRanking
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SeasonId"
      requestBody:
        description: Information needed to update the rankings
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRankingRequest"
      responses:
        "200":
          description: Rankings updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "201":
          description: New rankings created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid input, invalid ranking criteria, or no team results found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized (missing or invalid token)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Season regulation or team results not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/rankings/{team_result_id}:
    post:
      tags:
        - ranking
      summary: Create a new ranking
      description: Creates a new ranking for a team result in a season
      operationId: createRanking
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TeamResultId"
      requestBody:
        description: Information needed to create a ranking
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRankingRequest"
      responses:
        "201":
          description: Ranking created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid input or ranking already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized (missing or invalid token)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Team result not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/rankings/{id}:
    delete:
      tags:
        - ranking
      summary: Delete a ranking
      description: Deletes an existing ranking by its ID
      operationId: deleteRanking
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RankingId"
      responses:
        "200":
          description: Ranking deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid ID format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized (missing or invalid token)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Ranking not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    SeasonId:
      name: season_id
      in: path
      required: true
      schema:
        type: string
      example: "67d0ec1a412c1edc00d2117c"
...
    TeamResultId:
      name: id
      in: path
      required: true
      schema:
        type: string
      example: "67d0ec1a412c1edc00d2117b"
    RankingId:
      name: id
      in: path
      required: true
      schema:
        type: string
      example: "67d0ec1a412c1edc00d2117b"
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: "error"
        message:
          type: string
          example: "Invalid email address"
      required:
        - status
        - message
    SuccessResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        message:
          type: string
          example: "Operation successful"
        data:
          type: object
          nullable: true
          example: null
      required:
        - status
        - message
...
    Ranking:
      type: object
      properties:
        _id:
          type: string
          example: "67d0ec1a412c1edc00d2117b"
        team_result_id:
          type: string
          example: "67d0ec1a412c1edc00d2117d"
        season_id:
          type: string
          example: "67d0ec1a412c1edc00d2117c"
        rank:
          type: number
          example: 1
        date:
          type: string
          format: date-time
          example: "2025-04-08T00:00:00Z"
        createdAt:
          type: string
          format: date-time
          example: "2025-04-08T12:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-04-08T12:00:00Z"
      required:
        - _id
        - team_result_id
        - season_id
        - rank
        - date
        - createdAt
        - updatedAt
    RankingsResponse:
      type: object
      properties:
        status:
          type: string
          example: "success"
        message:
          type: string
          example: "Fetched season rankings successfully"
        data:
          type: array
          items:
            $ref: "#/components/schemas/Ranking"
      required:
        - status
        - message
        - data
    UpdateRankingRequest:
      type: object
      properties:
        match_date:
          type: string
          format: date-time
          example: "2025-04-08T12:00:00Z"
      required:
        - match_date
    CreateRankingRequest:
      type: object
      properties:
        season_id:
          type: string
          example: "67d0ec1a412c1edc00d2117c"
      required:
        - season_id
...